avalonia upgrade wip

+++ RogueEssence.Editor.Avalonia/DataEditor/Editors/Primitive/DoubleEditor.cs
@@ -31,7 +31,7 @@ namespace RogueEssence.Dev
                 nudValue.Minimum = attribute.Min;
                 nudValue.Maximum = attribute.Max;
             }
-            nudValue.Value = (double)member;
+            nudValue.Value = (decimal)member;
             control.Children.Add(nudValue);
         }
